+++ local.js
@@ -1144,37 +1144,14 @@ local.makePeerDomain = function makePeerDomain(user, relay, app, sid) {
 };
 
 // EXPORTED
-// breaks a proxy URI into the different parts
-// eg 'httpl://0.page/httpl%3A%2F%2Ffoo%2Fhttpl%253A%252F%252Fmy_worker.js%252F'
-// -> ['httpl://0.page/', 'httpl://foo/', 'httpl://my_worker.js/']
-local.parseProxyUri = function(uri) {
-	var parts = [];
-	var re = /^http(l|s)?/;
-	while (re.exec(uri)) {
-		var end = uri.indexOf('http', 8); // skip 8 to skip 'http(l|s)://'. If just http://, will skip 1 too many, but that shouldnt matter
-		if (end == -1)
-			break;
-		parts.push(uri.slice(0, end));
-		uri = decodeURIComponent(uri.slice(end));
-	}
-	parts.push(uri);
-	return parts;
-};
-
-// EXPORTED
-// builds a proxy URI out of an array of parts
-// eg ['httpl://0.page/', 'httpl://foo/', 'httpl://my_worker.js/']
-// -> 'httpl://0.page/httpl%3A%2F%2Ffoo%2Fhttpl%253A%252F%252Fmy_worker.js%252F'
-local.makeProxyUri = function(parts) {
-	var uri = 0;
-	for (var i=parts.length-1; i >= 0; i--) {
-		var part = parts[i];
-		if (part.hostname && part.proto) { // parsed via header
-			var proto = part.proto.name.toLowerCase();
-			part = ((proto&&(i===0||proto!='httpl'))?(proto+'://'):'') + part.hostname;
-		}
-		if (!uri) uri = part;
-		else uri = local.joinUri(part, encodeURIComponent(uri));
+// builds a proxy URI out of an array of templates
+// eg ('httpl://my_worker.js/', ['httpl://0.page/{uri}', 'httpl://foo/{?uri}'])
+// -> "httpl://0.page/httpl%3A%2F%2Ffoo%2F%3Furi%3Dhttpl%253A%252F%252Fmy_worker.js%252F"
+local.makeProxyUri = function(uri, templates) {
+	if (!Array.isArray(templates)) templates = [templates];
+	for (var i=templates.length-1; i >= 0; i--) {
+		var tmpl = templates[i];
+		uri = local.UriTemplate.parse(tmpl).expand({ uri: uri });
 	}
 	return uri;
 };
@@ -5210,7 +5187,7 @@ function Agent(context, parentAgent) {
 	this.context         = context         || null;
 	this.parentAgent = parentAgent || null;
 	this.links           = null;
-	// this.via             = null;
+	this.proxyTmpl       = null;
 	this.requestDefaults = null;
 }
 local.Agent = Agent;
@@ -5271,7 +5248,7 @@ Agent.prototype.dispatch = function(req) {
 			self.context.setResolved();
 			if (res.parsedHeaders.link) self.links = res.parsedHeaders.link;
 			else self.links = self.links || []; // cache an empty link list so we dont keep trying during resolution
-			// self.via = (res.parsedHeaders.via || null);
+			self.proxyTmpl = (res.header('Proxy-Tmpl')) ? res.header('Proxy-Tmpl').split(' ') : null;
 			return res;
 		})
 		.fail(function(res) {
@@ -5336,7 +5313,7 @@ Agent.prototype.follow = function(query) {
 Agent.prototype.unresolve = function() {
 	this.context.resetResolvedState();
 	this.links = null;
-	// this.via = null;
+	this.proxyTmpl = null;
 	return this;
 };
 
@@ -5436,8 +5413,8 @@ Agent.prototype.lookupLink = function(context) {
 			var link = local.queryLinks(this.links, context.query)[0];
 			if (link) {
 				var uri = local.UriTemplate.parse(link.href).expand(context.query);
-				// if (this.via)
-					// uri = local.makeProxyUri(this.via.concat(uri));
+				if (this.proxyTmpl && !link.noproxy)
+					uri = local.makeProxyUri(uri, this.proxyTmpl);
 				return uri;
 			}
 		}